Remove 'curl' and 'native' drivers.
