+++ config/recaptcha.php
@@ -28,17 +28,6 @@ return [
 
     /*
     |--------------------------------------------------------------------------
-    | Driver
-    |--------------------------------------------------------------------------
-    |
-    | Determine the http-client class.
-    | Only applies to v2/v3.
-    |
-    */
-    'driver'      => \OSMAviation\Recaptcha\Tools\Driver\Guzzle::class,
-
-    /*
-    |--------------------------------------------------------------------------
     | Options
     |--------------------------------------------------------------------------
     |
@@ -47,7 +36,6 @@ return [
     */
     'options'     => [
         'timeout' => 1,
-        'curl_verify' => true,
     ],
 
     /*
